+++ packages/browser/lib/ui/ApplePay/utilities.ts
@@ -4,6 +4,7 @@ import {
   PaymentTransactionDetails,
   RecurringTransactionDetails,
   TransactionDetailsWithDomain,
+  TransactionLineItem,
 } from "types";
 import {
   DisbursementContactAddress,
@@ -11,6 +12,7 @@ import {
   ValidateMerchantResponse,
   ApplePayCardNetwork,
   ApplePayPaymentRequest,
+  ShippingAddress,
 } from "./types";
 import ApplePayButton from ".";
 
@@ -27,6 +29,13 @@ type BuildSessionOptions = {
   disbursementOverrides?: {
     disbursementDetails?: PaymentDetailsInit;
   };
+  onShippingAddressChange?: (
+    newAddress: ShippingAddress
+  ) => Promise<{ amount: number; lineItems?: TransactionLineItem[] }>;
+  prepareTransaction?: () => Promise<{
+    amount?: number;
+    lineItems?: TransactionLineItem[];
+  }>;
 };
 
 export async function buildSession(
@@ -36,6 +45,9 @@ export async function buildSession(
   const { transaction: tx } = config;
 
   const merchant = await getMerchant(applePay, tx.merchantId);
+  if (!merchant) {
+    throw new Error("Merchant not found");
+  }
 
   let baseRequest: ApplePayPaymentRequest;
   if (tx.type === "payment") {
@@ -56,14 +68,63 @@ export async function buildSession(
   };
 
   // Apple pay requires subscribing to onshippingaddresschange and calling updateWith
-  // in order to receive shipping data. In future we should allow users to hook into
-  // this event to modify the payment request based on the shipping address.
+  // in order to receive shipping data.
   baseRequest.onshippingaddresschange = (event: PaymentRequestUpdateEvent) => {
-    const update: PaymentDetailsUpdate = {};
-    event.updateWith(update);
+    const target = event.target as unknown as {
+      shippingAddress?: ShippingAddress;
+    };
+    if (!target.shippingAddress) {
+      return;
+    }
+
+    if (config.onShippingAddressChange) {
+      const updates = updatePaymentRequest(
+        target.shippingAddress,
+        config,
+        tx,
+        merchant
+      ); // Do not await this promise
+      event.updateWith(updates);
+    } else {
+      // If no handler is provided, just update with empty shipping options
+      const update: PaymentDetailsUpdate = {};
+      event.updateWith(update);
+    }
   };
 
   return baseRequest;
+}
+
+async function updatePaymentRequest(
+  newAddress: ShippingAddress,
+  config: BuildSessionOptions,
+  tx: TransactionDetailsWithDomain,
+  merchant: MerchantDetail
+): Promise<PaymentDetailsUpdate> {
+  const updatedTransactionConfig = await config.onShippingAddressChange!(
+    newAddress
+  );
+  const displayItems = (updatedTransactionConfig.lineItems ?? []).map(
+    (item) => ({
+      label: item.label,
+      amount: {
+        value: (item.amount / 100).toFixed(2).toString(),
+        currency: tx.currency,
+      },
+    })
+  );
+  const total = {
+    label: merchant.name,
+    amount: {
+      currency: tx.currency,
+      value: (updatedTransactionConfig.amount / 100).toFixed(2),
+    },
+  };
+  const updates = {
+    displayItems,
+    total,
+  };
+  return updates;
 }
 
 function buildPaymentSession(
@@ -110,6 +171,7 @@ function buildPaymentSession(
     requestPayerPhone: config.requestPayerDetails?.includes("phone") ?? false,
     requestShipping: config.requestShipping ?? false,
     shippingType: "shipping",
+    onShippingAddressChange: config.onShippingAddressChange,
   };
 
   const paymentOverrides = config.paymentOverrides || {};
@@ -355,7 +417,10 @@ async function validateMerchant(
   return response.json();
 }
 
-async function getMerchant(applePay: ApplePayButton, id: string) {
+async function getMerchant(
+  applePay: ApplePayButton,
+  id: string
+): Promise<MerchantDetail | undefined> {
   const app = applePay.client.config.appId;
   const apiURL = applePay.client.config.http.apiUrl;
   const response = await fetch(`${apiURL}/frontend/merchants/${id}`, {