+++ examples/three-d-secure/CHANGELOG.md
@@ -1,5 +1,17 @@
 # example-three-d-secure
 
+## 0.0.39
+
+### Patch Changes
+
+- Updated dependencies [df76034]
+- Updated dependencies [31c1ac8]
+- Updated dependencies [e503d31]
+- Updated dependencies [2640be6]
+- Updated dependencies [cd08b80]
+- Updated dependencies [8dec7e4]
+  - @evervault/js@2.2.0
+
 ## 0.0.38
 
 ### Patch Changes