+++ packages/inputs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @evervault/inputs
 
+## 2.18.17
+
+### Patch Changes
+
+- Updated dependencies [df76034]
+- Updated dependencies [c4db698]
+- Updated dependencies [31c1ac8]
+- Updated dependencies [e503d31]
+- Updated dependencies [cd08b80]
+- Updated dependencies [8dec7e4]
+  - @evervault/browser@2.46.0
+
 ## 2.18.16
 
 ### Patch Changes